В папку сделал
