+++ src/main/java/Message/HelpMessage.java
@@ -1,3 +1,5 @@
+package Message;
+
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageText;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
@@ -7,13 +9,13 @@ import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKe
 import java.util.ArrayList;
 import java.util.List;
 
-public class MenuMessage {
-    String mainMenuText = "✋ Добро пожаловать в главное меню бота Mix & Cook! Для просмотра рецептов воспользуйтесь кнопками ниже:";
+public class HelpMessage {
+    String HelpText = " Это бот для приготовление всяких блюд.\n" + "\n" + "Информация по всем функциям бота доступна по команде /start\n" + "\n" + "ℹ Подробная инструкция будет потом\n";
 
     public SendMessage createMessage(long chatId) {
         return SendMessage.builder()
                 .chatId(chatId)
-                .text(mainMenuText)
+                .text(HelpText)
                 .replyMarkup(createKeyboard())
                 .build();
     }
@@ -22,19 +24,15 @@ public class MenuMessage {
         return EditMessageText.builder()
                 .chatId(String.valueOf(chatId))
                 .messageId(messageId)
-                .text(mainMenuText)
+                .text(HelpText)
                 .replyMarkup(createKeyboard())
                 .build();
     }
 
-    private  InlineKeyboardMarkup createKeyboard() {
+    private InlineKeyboardMarkup createKeyboard() {
         List<InlineKeyboardRow> rows = new ArrayList<>();
 
-        rows.add(createRow("Поиск блюд, напитков", "/search"));
-        rows.add(createRow("Случайный рецепт", "/random"));
-        rows.add(createRow("Избранное", "/wishlist"));
-        rows.add(createRow("Каталог", "/catalog"));
-        rows.add(createRow("Справка", "/help"));
+        rows.add(createRow("На главное меню", "/back"));
 
         return InlineKeyboardMarkup.builder().keyboard(rows).build();
     }
